add:ユーザー検証を追加 #35

+++ back/app/controllers/api/v1/item_likes_controller.rb
@@ -1,5 +1,6 @@
 class Api::V1::ItemLikesController < ApplicationController
   def create
+    return if User.find(params[:user_id]).email != params[:uid]
     like = ItemLike.new(item_like_params.except(:uid))
     if like.save
       item_likes = ItemLike.filter_by_item(params[:item_id]).select(:id, :user_id, :item_id)
@@ -9,6 +10,9 @@ class Api::V1::ItemLikesController < ApplicationController
 
   def destroy
     like = ItemLike.find_by(user_id: params[:user_id], item_id: params[:item_id])
+
+    return if User.find(like.user_id).email != params[:uid]
+
     if like.destroy
       item_likes = ItemLike.filter_by_item(params[:item_id]).select(:id, :user_id, :item_id)
       render json: item_likes.as_json(except: [:created_at, :updated_at])
@@ -17,6 +21,6 @@ class Api::V1::ItemLikesController < ApplicationController
 
   private
     def item_like_params
-      params.permit(:user_id, :item_id)
+      params.permit(:user_id, :item_id, :uid)
     end
 end
